requirements.txt
streamlit==1.39.0
openai==1.51.2
python-dotenv==1.0.1
